Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 
 int main() {
-    int n,power = 1;
+    long int n,power = 1;
     scanf("%d",&n);
-    int result = 0;
+    long int result = 0;
     while(n){
         result = result + ((n&1)*power);
         power = power * 10;
         n = n>>1;
     }
-    printf("%d",result);
+    printf("%ld",result);
     return 0;
 }
